+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All package updates & migration steps will be listed in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.0] - 2024-04-30
+### Added
+- Support code for building the parameters runtime outside of Unity
+- `FilePathAssetLoader` for non Unity loading
+- Method `IBaseInfo Get(string identifier, Type type)` to the `ParameterManager` and `Params`
+
 ## [3.5.0] - 2024-02-21
 ### Added
 - Support for displaying only the correct sub class of ParameterScriptableObjects in Inspector